Rewrite colors by hand
With treesitter coming to Neovim in 0.5 I thought it time to get back at this theme. I've scrapped the Textmate conversion as I had to override it too often and instead taken to matching variables by hand from the VSCode theme. This isn't a 1:1 match yet, but I expect most of the work is done with language-specific changes coming in patches.